写的很糟糕，自己都有些乱

+++ Spider.class.php
@@ -1,6 +1,6 @@
 <?php 
 /**
-* 
+* CURLOPT_RETURNTRANSFER
 */
 class Spider
 {
@@ -8,36 +8,53 @@ class Spider
 	protected $pattern='/\<a.*?href=.*?\>/';
 	protected $url_file='url.txt';
 	protected $web_content_dir='../webContent/';
-	public function __construct($url)
+	protected $ch=null;
+
+	public function __construct($url='')
 	{
-		$this->url=$url;
+		$this->ch=curl_init();
+		if (!empty($url)) {
+			$this->url=$url;
+		}
 	}
 
 	public function crawlerWbeContent(){
 		$url=file_get_contents($this->url_file);
 		$url=explode("\r\n", $url);
 		foreach ($url as $k => $v) {
-			$tmp=file_get_contents($v);
-			file_put_contents($this->web_content_dir.$k, $tmp);
+			curl_setopt($this->ch, CURLOPT_URL, $v);
+			curl_setopt($this->ch, CURLOPT_RETURNTRANSFER, true);
+			$tmp=curl_exec($this->ch);
+			if (file_exists($this->web_content_dir.$k)) {
+				file_put_contents($this->web_content_dir.$k, $tmp, FILE_APPEND);
+			}else{
+				file_put_contents($this->web_content_dir.$k, $tmp);
+			}	
 			$this->_writeUrl($tmp);
+		}
+		$this->_close();
+	}
+
+	public function getUrlFile($pattern='/href=\".*?\"/',$scheme='http',$start=6){
+		curl_setopt($this->ch, CURLOPT_URL, $this->url);
+		curl_setopt($this->ch, CURLOPT_RETURNTRANSFER, true);
+		$str=curl_exec($this->ch);
+		$this->_writeUrl($str,$pattern,$scheme,$start);
+	}
+
+	public function __call($name,$argument){
+		if (stripos($name, 'set_')==0) {
+			$variable=substr($name, 4);
+			if (isset($this->$variable)) {
+				$this->$variable=$argument[0];
+			}else{
+				echo 'Invalid variable';
+			}	
 		}
 	}
 
-	public function getUrlFile(){
-		$str=file_get_contents($this->url);
-		$this->_writeUrl($str);
-	}
-
-	public function setUrlFile($string){
-		$this->url_file=$string;
-	}
-
-	public function setUrl($string){
-		$this->url=$string;
-	}
-
-	protected function _writeUrl($str){
-		$tmp=$this->_filterUrl($str);
+	protected function _writeUrl($str,$pattern='/href=\".*?\"/',$scheme='http',$start=6){
+		$tmp=$this->_filterUrl($str,$pattern,$scheme,$start);
 		if (file_exists($this->url_file)) {
 			file_put_contents($this->url_file, $tmp, FILE_APPEND);
 		}else{
@@ -45,14 +62,20 @@ class Spider
 		}
 	}
 
-	protected function _filterUrl($string){
+	protected function _filterUrl($string,$pattern='/href=\".*?\"/',$scheme='http',$start=6,$is_fill_url=false){
 		preg_match_all($this->pattern, $string, $arr);
-		$pattern='/href=\".*?\"/';
 		$url=array();
 		foreach ($arr[0] as $k => $v) {
 			if(preg_match($pattern, $v, $tmp)){
-				if (stripos($tmp[0], 'http')) {
-					$url[]=substr($tmp[0], 6, -1);
+				if (stripos($tmp[0], $scheme)) {
+					$tmp=substr($tmp[0], $start, -1);
+					if ($is_fill_url) {
+						$host=curl_getinfo($this->ch,CURLINFO_EFFECTIVE_URL);
+						$tmp=parse_url($host,PHP_URL_HOST).$tmp;
+					}
+					if (!empty($tmp)) {
+						$url[]=$tmp;
+					}
 				}
 			}
 		}
@@ -60,6 +83,10 @@ class Spider
 		$url=array_filter($url);
 		$url=implode("\r\n", $url);
 		return $url;
+	}
+
+	protected function _close(){
+		curl_close($this->ch);
 	}
 }
 ?>
